--- jeecg-boot/jeecg-cloud-module/jeecg-cloud-system-start/src/main/java/org/jeecg/modules/cloud/feign/feign/JeecgTestClientDyn.java
@@ -1,14 +1,0 @@
-package org.jeecg.modules.cloud.feign.feign;
-
-import org.jeecg.common.api.vo.Result;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-
-/**
- * 动态feign接口定义
- */
-public interface JeecgTestClientDyn {
-
-    @GetMapping(value = "/test/getMessage")
-    Result<String> getMessage(@RequestParam("name") String name);
-}
