--- jeecg-boot/jeecg-cloud-module/jeecg-cloud-system-start/src/main/java/org/jeecg/modules/cloud/feign/feign/fallback/JeecgTestClientFallback.java
@@ -1,17 +1,0 @@
-package org.jeecg.modules.cloud.feign.feign.fallback;
-
-import feign.hystrix.FallbackFactory;
-import org.jeecg.modules.cloud.feign.feign.JeecgTestClient;
-import org.springframework.stereotype.Component;
-
-/**
- * @author qinfeng
- */
-@Component
-public class JeecgTestClientFallback implements FallbackFactory<JeecgTestClient> {
-
-    @Override
-    public JeecgTestClient create(Throwable throwable) {
-        return null;
-    }
-}
