+++ jeecg-boot/jeecg-cloud-module/jeecg-cloud-test/jeecg-cloud-test-more/src/main/java/org/jeecg/modules/test/feign/factory/JeecgTestClientFactory.java
@@ -1,0 +1,23 @@
+package org.jeecg.modules.test.feign.factory;
+
+
+
+
+import feign.hystrix.FallbackFactory;
+import org.jeecg.modules.test.feign.client.JeecgTestClient;
+import org.jeecg.modules.test.feign.fallback.JeecgTestFallback;
+import org.springframework.stereotype.Component;
+
+/**
+ * @author qinfeng
+ */
+@Component
+public class JeecgTestClientFactory implements FallbackFactory<JeecgTestClient> {
+
+    @Override
+    public JeecgTestClient create(Throwable throwable) {
+        JeecgTestFallback fallback = new JeecgTestFallback();
+        fallback.setCause(throwable);
+        return fallback;
+    }
+}
