+++ jeecg-boot/jeecg-cloud-module/jeecg-cloud-test/jeecg-cloud-test-more/src/main/java/org/jeecg/modules/test/xxljob/DemoJobHandler.java
@@ -1,0 +1,234 @@
+
+package org.jeecg.modules.test.xxljob;
+
+
+import com.xxl.job.core.biz.model.ReturnT;
+import com.xxl.job.core.handler.IJobHandler;
+import com.xxl.job.core.handler.annotation.XxlJob;
+import com.xxl.job.core.log.XxlJobLogger;
+import com.xxl.job.core.util.ShardingUtil;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import java.io.BufferedInputStream;
+import java.io.BufferedReader;
+import java.io.DataOutputStream;
+import java.io.InputStreamReader;
+import java.net.HttpURLConnection;
+import java.net.URL;
+import java.util.Arrays;
+
+
+/**
+ * xxl-job定时任务测试
+ */
+@Component
+@Slf4j
+public class DemoJobHandler {
+
+
+    /**
+     * 简单任务
+     *
+     * @param params
+     * @return
+     */
+
+    @XxlJob(value = "demoJob")
+    public ReturnT<String> demoJobHandler(String params) {
+        log.info("我是定时任务,我执行了...............................");
+        return ReturnT.SUCCESS;
+    }
+
+    /**
+     * 2、分片广播任务
+     */
+
+    @XxlJob("shardingJobHandler")
+    public ReturnT<String> shardingJobHandler(String param) throws Exception {
+
+        // 分片参数
+        ShardingUtil.ShardingVO shardingVO = ShardingUtil.getShardingVo();
+        XxlJobLogger.log("分片参数：当前分片序号 = {}, 总分片数 = {}", shardingVO.getIndex(), shardingVO.getTotal());
+
+        // 业务逻辑
+        for (int i = 0; i < shardingVO.getTotal(); i++) {
+            if (i == shardingVO.getIndex()) {
+                XxlJobLogger.log("第 {} 片, 命中分片开始处理", i);
+            } else {
+                XxlJobLogger.log("第 {} 片, 忽略", i);
+            }
+        }
+
+        return ReturnT.SUCCESS;
+    }
+
+
+    /**
+     * 3、命令行任务
+     */
+
+    @XxlJob("commandJobHandler")
+    public ReturnT<String> commandJobHandler(String param) throws Exception {
+        String command = param;
+        int exitValue = -1;
+
+        BufferedReader bufferedReader = null;
+        try {
+            // command process
+            Process process = Runtime.getRuntime().exec(command);
+            BufferedInputStream bufferedInputStream = new BufferedInputStream(process.getInputStream());
+            bufferedReader = new BufferedReader(new InputStreamReader(bufferedInputStream));
+
+            // command log
+            String line;
+            while ((line = bufferedReader.readLine()) != null) {
+                XxlJobLogger.log(line);
+            }
+
+            // command exit
+            process.waitFor();
+            exitValue = process.exitValue();
+        } catch (Exception e) {
+            XxlJobLogger.log(e);
+        } finally {
+            if (bufferedReader != null) {
+                bufferedReader.close();
+            }
+        }
+
+        if (exitValue == 0) {
+            return IJobHandler.SUCCESS;
+        } else {
+            return new ReturnT<String>(IJobHandler.FAIL.getCode(), "command exit value(" + exitValue + ") is failed");
+        }
+    }
+
+
+    /**
+     * 4、跨平台Http任务
+     * 参数示例：
+     * "url: http://www.baidu.com\n" +
+     * "method: get\n" +
+     * "data: content\n";
+     */
+
+    @XxlJob("httpJobHandler")
+    public ReturnT<String> httpJobHandler(String param) throws Exception {
+
+        // param parse
+        if (param == null || param.trim().length() == 0) {
+            XxlJobLogger.log("param[" + param + "] invalid.");
+            return ReturnT.FAIL;
+        }
+        String[] httpParams = param.split("\n");
+        String url = null;
+        String method = null;
+        String data = null;
+        for (String httpParam : httpParams) {
+            if (httpParam.startsWith("url:")) {
+                url = httpParam.substring(httpParam.indexOf("url:") + 4).trim();
+            }
+            if (httpParam.startsWith("method:")) {
+                method = httpParam.substring(httpParam.indexOf("method:") + 7).trim().toUpperCase();
+            }
+            if (httpParam.startsWith("data:")) {
+                data = httpParam.substring(httpParam.indexOf("data:") + 5).trim();
+            }
+        }
+
+        // param valid
+        if (url == null || url.trim().length() == 0) {
+            XxlJobLogger.log("url[" + url + "] invalid.");
+            return ReturnT.FAIL;
+        }
+        if (method == null || !Arrays.asList("GET", "POST").contains(method)) {
+            XxlJobLogger.log("method[" + method + "] invalid.");
+            return ReturnT.FAIL;
+        }
+
+        // request
+        HttpURLConnection connection = null;
+        BufferedReader bufferedReader = null;
+        try {
+            // connection
+            URL realUrl = new URL(url);
+            connection = (HttpURLConnection) realUrl.openConnection();
+
+            // connection setting
+            connection.setRequestMethod(method);
+            connection.setDoOutput(true);
+            connection.setDoInput(true);
+            connection.setUseCaches(false);
+            connection.setReadTimeout(5 * 1000);
+            connection.setConnectTimeout(3 * 1000);
+            connection.setRequestProperty("connection", "Keep-Alive");
+            connection.setRequestProperty("Content-Type", "application/json;charset=UTF-8");
+            connection.setRequestProperty("Accept-Charset", "application/json;charset=UTF-8");
+
+            // do connection
+            connection.connect();
+
+            // data
+            if (data != null && data.trim().length() > 0) {
+                DataOutputStream dataOutputStream = new DataOutputStream(connection.getOutputStream());
+                dataOutputStream.write(data.getBytes("UTF-8"));
+                dataOutputStream.flush();
+                dataOutputStream.close();
+            }
+
+            // valid StatusCode
+            int statusCode = connection.getResponseCode();
+            if (statusCode != 200) {
+                throw new RuntimeException("Http Request StatusCode(" + statusCode + ") Invalid.");
+            }
+
+            // result
+            bufferedReader = new BufferedReader(new InputStreamReader(connection.getInputStream(), "UTF-8"));
+            StringBuilder result = new StringBuilder();
+            String line;
+            while ((line = bufferedReader.readLine()) != null) {
+                result.append(line);
+            }
+            String responseMsg = result.toString();
+
+            XxlJobLogger.log(responseMsg);
+            return ReturnT.SUCCESS;
+        } catch (Exception e) {
+            XxlJobLogger.log(e);
+            return ReturnT.FAIL;
+        } finally {
+            try {
+                if (bufferedReader != null) {
+                    bufferedReader.close();
+                }
+                if (connection != null) {
+                    connection.disconnect();
+                }
+            } catch (Exception e2) {
+                XxlJobLogger.log(e2);
+            }
+        }
+
+    }
+
+
+    /**
+     * 5、生命周期任务示例：任务初始化与销毁时，支持自定义相关逻辑；
+     */
+
+    @XxlJob(value = "demoJobHandler2", init = "init", destroy = "destroy")
+    public ReturnT<String> demoJobHandler2(String param) throws Exception {
+        XxlJobLogger.log("XXL-JOB, Hello World.");
+        return ReturnT.SUCCESS;
+    }
+
+    public void init() {
+        log.info("init");
+    }
+
+    public void destroy() {
+        log.info("destory");
+    }
+
+}
+
