+++ jeecg-boot/jeecg-cloud-module/jeecg-cloud-test/jeecg-cloud-test-rabbitmq/pom.xml
@@ -3,26 +3,20 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>jeecg-cloud-module</artifactId>
+        <artifactId>jeecg-cloud-test</artifactId>
         <groupId>org.jeecgframework.boot</groupId>
         <version>3.1.0</version>
     </parent>
-
     <modelVersion>4.0.0</modelVersion>
-    <packaging>pom</packaging>
-    <artifactId>jeecg-cloud-test</artifactId>
+    <description>消息队列测试模块</description>
+    <artifactId>jeecg-cloud-test-rabbitmq</artifactId>
 
     <dependencies>
+        <!--rabbitmq消息队列-->
         <dependency>
             <groupId>org.jeecgframework.boot</groupId>
-            <artifactId>jeecg-boot-base-core</artifactId>
+            <artifactId>jeecg-boot-starter-rabbitmq</artifactId>
         </dependency>
     </dependencies>
 
-    <modules>
-        <module>jeecg-cloud-test-seata</module>
-        <module>jeecg-cloud-test-shardingsphere</module>
-        <module>jeecg-cloud-test-more</module>
-        <module>jeecg-cloud-test-rabbitmq</module>
-    </modules>
 </project>
