+++ jeecg-boot/jeecg-cloud-module/jeecg-cloud-test/jeecg-cloud-test-rabbitmq/src/main/java/org/jeecg/modules/test/rabbitmq/listener/HelloTimeReceiver.java
@@ -1,0 +1,34 @@
+package org.jeecg.modules.test.rabbitmq.listener;
+
+import org.jeecg.boot.starter.rabbitmq.core.BaseRabbiMqHandler;
+import org.jeecg.boot.starter.rabbitmq.listenter.MqListener;
+import org.jeecg.common.annotation.RabbitComponent;
+import org.jeecg.common.base.BaseMap;
+import org.jeecg.modules.test.rabbitmq.constant.CloudConstant;
+import org.springframework.amqp.rabbit.annotation.RabbitHandler;
+import org.springframework.amqp.rabbit.annotation.RabbitListener;
+import org.springframework.amqp.support.AmqpHeaders;
+import org.springframework.messaging.handler.annotation.Header;
+
+import com.rabbitmq.client.Channel;
+
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@RabbitListener(queues = CloudConstant.MQ_JEECG_PLACE_ORDER_TIME)
+@RabbitComponent(value = "helloTimeReceiver")
+public class HelloTimeReceiver extends BaseRabbiMqHandler<BaseMap> {
+
+    @RabbitHandler
+    public void onMessage(BaseMap baseMap, Channel channel, @Header(AmqpHeaders.DELIVERY_TAG) long deliveryTag) {
+        super.onMessage(baseMap, deliveryTag, channel, new MqListener<BaseMap>() {
+            @Override
+            public void handler(BaseMap map, Channel channel) {
+                //业务处理
+                String orderId = map.get("orderId").toString();
+                log.info("Time Receiver1，orderId : " + orderId);
+            }
+        });
+    }
+
+}
